More refactoring. Converting @context aspect to be a network attribute. Remove the ProvenanceHistory aspect when converting networks to cx2.
